Use service factories for default locale providers
Certified spring moment
