+++ src/main/kotlin/io/github/freya022/botcommands/internal/localization/text/DefaultTextCommandLocaleProvider.kt
@@ -10,24 +10,36 @@ import io.github.freya022.botcommands.api.localization.text.TextCommandLocalePro
 import io.github.freya022.botcommands.internal.utils.classRef
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
 import java.util.*
 
+// spring moment
+// need to make a (totally useless) factory for a simple object,
+// as ConditionalOnMissingBean doesn't always work
 @BService
-@ConditionalService(DefaultTextCommandLocaleProvider.ExistingSupplierChecker::class)
-@ConditionalOnMissingBean(TextCommandLocaleProvider::class)
-internal class DefaultTextCommandLocaleProvider : TextCommandLocaleProvider {
-    override fun getLocale(event: MessageReceivedEvent): Locale = event.guild.locale.toLocale()
+@Configuration
+internal open class DefaultTextCommandLocaleProviderProvider internal constructor() {
+    private object DefaultTextCommandLocaleProvider : TextCommandLocaleProvider {
+        override fun getLocale(event: MessageReceivedEvent): Locale = event.guild.locale.toLocale()
 
-    internal object ExistingSupplierChecker : ConditionalServiceChecker {
-        override fun checkServiceAvailability(serviceContainer: ServiceContainer, checkedClass: Class<*>): String? {
-            // Try to get TextCommandLocaleProvider interfaced services, except ours
-            // If empty, then the user didn't provide one, in which case we can allow
-            //Won't take DefaultTextCommandLocaleProvider into account
-            val providers = serviceContainer.getInterfacedServices<TextCommandLocaleProvider>()
-            if (providers.isNotEmpty())
-                return "An user supplied ${classRef<TextCommandLocaleProvider>()} is already active (${providers.first().javaClass.simpleNestedName})"
+        object ExistingSupplierChecker : ConditionalServiceChecker {
+            override fun checkServiceAvailability(serviceContainer: ServiceContainer, checkedClass: Class<*>): String? {
+                // Try to get TextCommandLocaleProvider interfaced services, except ours
+                // If empty, then the user didn't provide one, in which case we can allow
+                //Won't take DefaultTextCommandLocaleProvider into account
+                val providers = serviceContainer.getInterfacedServices<TextCommandLocaleProvider>()
+                if (providers.isNotEmpty())
+                    return "An user supplied ${classRef<TextCommandLocaleProvider>()} is already active (${providers.first().javaClass.simpleNestedName})"
 
-            return null
+                return null
+            }
         }
     }
+
+    @BService
+    @ConditionalService(DefaultTextCommandLocaleProvider.ExistingSupplierChecker::class)
+    @Bean
+    @ConditionalOnMissingBean(TextCommandLocaleProvider::class)
+    open fun defaultTextCommandLocaleProvider(): TextCommandLocaleProvider = DefaultTextCommandLocaleProvider
 }
